backend/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
sqlalchemy
psycopg2
python-jose
python-multipart